Add HostAccessor and SystemRequest
HostAccessor is a provider/supplier of HostControlSystem instances that
works for both local and remote hosts.

Remove HostControlSystems and refactor local and SSH systems to work
with the new model.

Temporarily reduce code coverage limits in giraffe-core until the
coverage system can be fixed (e.g. focus on lines, statements, and
branches only).