add option to specify keepalive interval for ssh connections
Note that by default, sshj uses an old "heartbeat" message for
keepalives, which is a standard ssh packet. It's actually possible to
configure it to instead use ssh protocol keepalive messages, but doing
so requires setting the keepalive provider in the Config objecct that's
passed to SshConnectorFactory. That's buried in some abstraction, so
this change doesn't actually allow turning on real ssh keepalives
directly, and instead uses the default heartbeating method.